Add define_lang_parser macro and use in parsers
