+++ core/src/parser/java.rs
@@ -1,11 +1,1 @@
-use tree_sitter::{Language, Parser, Tree};
-
-pub fn language() -> Language {
-    tree_sitter_java::LANGUAGE.into()
-}
-
-pub fn parse(source: &str, old_tree: Option<&Tree>) -> Option<Tree> {
-    let mut parser = Parser::new();
-    parser.set_language(&language()).ok()?;
-    parser.parse(source, old_tree)
-}
+crate::define_lang_parser!(tree_sitter_java::LANGUAGE);
